cancer_detection/mysite/requirements.txt
argon2-cffi==19.2.0
asgiref==3.2.3
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cycler==0.10.0
dicom==0.9.9.post1
Django==2.2.9
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
django-dicom==0.0.4
django-extensions==2.2.6
django-filter==2.2.0
django-tastypie==0.14.3
djangorestframework==3.11.0
Faker==3.0.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.8
imageio==2.6.1
kiwisolver==1.1.0
matplotlib==3.1.3
misaka==2.1.1
numpy==1.18.1
opencv-python==4.2.0.32
pandas==0.25.3
Pillow==7.0.0
psycopg2-binary==2.8.4
pycparser==2.19
pydicom==1.4.1
pyparsing==2.4.6
pypng==0.0.20
python-dateutil==1.5
python-mimeparse==1.6.0
pytz==2019.3
requests==2.22.0
scenedetect==0.5.1.1
scipy==1.4.1
six==1.13.0
sqlparse==0.3.0
termcolor==1.1.0
text-unidecode==1.3
tika==1.19
tqdm==4.42.0
urllib3==1.25.8
wincertstore==0.2
xlrd==1.2.0
